+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-spotify-wrapper` will be documented in this file.
 
+## 0.1.4 - 2018-11-19
+
+- Add "Request Access Token" method.
+- Add `show_dialog` option.
+
 ## 0.1.3 - 2018-11-15
 
 - Fix some issues.